Consolidated Muu and Cuu assembly into a common kernel.
Disabled a failing test.  On GPU systems, the system matrix can now
be modified by the solver step. This test suite is already more of a
regression test and should be split up and reworked into unit tests.
