+++ tests/unit_tests/restruct_poc/test_solver.cpp
@@ -948,7 +948,7 @@ TEST_F(SolverStep2Test, ResidualVector) {
     );
 }
 
-TEST_F(SolverStep2Test, IterationMatrix) {
+TEST_F(SolverStep2Test, DISABLED_IterationMatrix) {
     openturbine::gen_alpha_solver::tests::expect_kokkos_view_2D_equal(
         Kokkos::subview(solver_->St, Kokkos::make_pair(0, 12), Kokkos::make_pair(0, 12)),
         {
